@@ -1,10 +1,6 @@
 module.exports = {
-    purge: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
-    darkMode: false, // or 'media' or 'class'
+    content: ['./src/**/*.{js,jsx,ts,tsx}', './public/index.html'],
     theme: {
-    extend: {},
-    },
-    variants: {
     extend: {},
     },
     plugins: [
